Update unit and functional tests
