+++ pkg/virt-controller/services/template_test.go
@@ -50,6 +50,7 @@ import (
 	"kubevirt.io/client-go/kubecli"
 	fakenetworkclient "kubevirt.io/client-go/networkattachmentdefinitionclient/fake"
 
+	"kubevirt.io/kubevirt/pkg/hypervisor"
 	"kubevirt.io/kubevirt/pkg/pointer"
 
 	k6tconfig "kubevirt.io/kubevirt/pkg/config"
@@ -500,6 +501,7 @@ var _ = Describe("Template", func() {
 					"--hook-sidecars", "1",
 					"--ovmf-path", ovmfPath,
 					"--disk-memory-limit", strconv.Itoa(virtconfig.DefaultDiskVerificationMemoryLimitBytes),
+					"--hypervisor", config.GetHypervisor().Name,
 				}))
 				Expect(pod.Spec.Containers[1].Name).To(Equal("hook-sidecar-0"))
 				Expect(pod.Spec.Containers[1].Image).To(Equal("some-image:v1"))
@@ -1098,6 +1100,7 @@ var _ = Describe("Template", func() {
 					"--hook-sidecars", "1",
 					"--ovmf-path", ovmfPath,
 					"--disk-memory-limit", strconv.Itoa(virtconfig.DefaultDiskVerificationMemoryLimitBytes),
+					"--hypervisor", config.GetHypervisor().Name,
 				}))
 				Expect(pod.Spec.Containers[1].Name).To(Equal("hook-sidecar-0"))
 				Expect(pod.Spec.Containers[1].Image).To(Equal("some-image:v1"))
@@ -2356,7 +2359,8 @@ var _ = Describe("Template", func() {
 						if vmiCPUArch == "" {
 							vmiCPUArch = svc.clusterConfig.GetClusterCPUArch()
 						}
-						overhead := GetMemoryOverhead(&vmi, vmiCPUArch, svc.clusterConfig.GetConfig().AdditionalGuestMemoryOverheadRatio)
+						launcherRenderer := hypervisor.NewLauncherResourceRenderer(config.GetHypervisor().Name)
+						overhead := launcherRenderer.GetMemoryOverhead(&vmi, vmiCPUArch, svc.clusterConfig.GetConfig().AdditionalGuestMemoryOverheadRatio)
 
 						mem100.Sub(overhead)
 						mem110.Sub(overhead)
@@ -2892,7 +2896,8 @@ var _ = Describe("Template", func() {
 				arch := config.GetClusterCPUArch()
 				Expect(err).ToNot(HaveOccurred())
 				expectedMemory := resource.NewScaledQuantity(0, resource.Kilo)
-				expectedMemory.Add(GetMemoryOverhead(vmi, arch, config.GetConfig().AdditionalGuestMemoryOverheadRatio))
+				launcherRenderer := hypervisor.NewLauncherResourceRenderer(config.GetHypervisor().Name)
+				expectedMemory.Add(launcherRenderer.GetMemoryOverhead(vmi, arch, config.GetConfig().AdditionalGuestMemoryOverheadRatio))
 				expectedMemory.Add(*vmi.Spec.Domain.Resources.Requests.Memory())
 				Expect(pod.Spec.Containers[0].Resources.Requests.Memory().Value()).To(Equal(expectedMemory.Value()))
 			})
@@ -2920,7 +2925,8 @@ var _ = Describe("Template", func() {
 				arch := config.GetClusterCPUArch()
 				Expect(err).ToNot(HaveOccurred())
 				expectedMemory := resource.NewScaledQuantity(0, resource.Kilo)
-				expectedMemory.Add(GetMemoryOverhead(vmi1, arch, config.GetConfig().AdditionalGuestMemoryOverheadRatio))
+				launcherRenderer := hypervisor.NewLauncherResourceRenderer(config.GetHypervisor().Name)
+				expectedMemory.Add(launcherRenderer.GetMemoryOverhead(vmi1, arch, config.GetConfig().AdditionalGuestMemoryOverheadRatio))
 				expectedMemory.Add(*vmi.Spec.Domain.Resources.Requests.Memory())
 				Expect(pod.Spec.Containers[0].Resources.Requests.Memory().Value()).To(Equal(expectedMemory.Value()))
 				Expect(pod1.Spec.Containers[0].Resources.Requests.Memory().Value()).To(Equal(expectedMemory.Value()))
@@ -3027,6 +3033,7 @@ var _ = Describe("Template", func() {
 
 		It("should call sidecar creators", func() {
 			config, _, _ := testutils.NewFakeClusterConfigUsingKVWithCPUArch(kv, defaultArch)
+
 			svc = NewTemplateService("kubevirt/virt-launcher",
 				240,
 				"/var/run/kubevirt",
@@ -4753,7 +4760,8 @@ var _ = Describe("Template", func() {
 				arch := config.GetClusterCPUArch()
 				Expect(err).ToNot(HaveOccurred())
 				expectedMemory := resource.NewScaledQuantity(0, resource.Kilo)
-				expectedMemory.Add(GetMemoryOverhead(vmi, arch, config.GetConfig().AdditionalGuestMemoryOverheadRatio))
+				launcherRenderer := hypervisor.NewLauncherResourceRenderer(config.GetHypervisor().Name)
+				expectedMemory.Add(launcherRenderer.GetMemoryOverhead(vmi, arch, config.GetConfig().AdditionalGuestMemoryOverheadRatio))
 				expectedMemory.Add(*vmi.Spec.Domain.Resources.Requests.Memory())
 				Expect(pod.Spec.Containers[0].Resources.Requests.Memory().Value()).To(Equal(expectedMemory.Value()))
 			})
@@ -4825,45 +4833,6 @@ var _ = Describe("Template", func() {
 			})
 		})
 
-		Context("with guest-to-request memory headroom", func() {
-			BeforeEach(func() {
-				config, kvStore, svc = configFactory(defaultArch)
-			})
-
-			newVmi := func() *v1.VirtualMachineInstance {
-				vmi := api.NewMinimalVMI("test-vmi")
-
-				vmi.Spec.Domain.Resources = v1.ResourceRequirements{
-					Requests: k8sv1.ResourceList{
-						k8sv1.ResourceMemory: resource.MustParse("1G"),
-						k8sv1.ResourceCPU:    resource.MustParse("1"),
-					},
-				}
-
-				return vmi
-			}
-
-			DescribeTable("should add guest-to-memory headroom when configured with ratio", func(ratioStr string) {
-				vmi := newVmi()
-
-				ratio, err := strconv.ParseFloat(ratioStr, 64)
-				Expect(err).ToNot(HaveOccurred())
-
-				originalOverhead := GetMemoryOverhead(vmi, config.GetClusterCPUArch(), nil)
-				actualOverheadWithHeadroom := GetMemoryOverhead(vmi, config.GetClusterCPUArch(), pointer.P(ratioStr))
-				expectedOverheadWithHeadroom := multiplyMemory(originalOverhead, ratio)
-
-				const errFmt = "overhead without headroom: %s, ratio: %s, actual overhead with headroom: %s, expected overhead with headroom: %s"
-				Expect(newVmi()).To(Equal(vmi), "vmi object should not be changed")
-				Expect(actualOverheadWithHeadroom.Cmp(expectedOverheadWithHeadroom)).To(Equal(0),
-					fmt.Sprintf(errFmt, originalOverhead.String(), ratioStr, actualOverheadWithHeadroom.String(), expectedOverheadWithHeadroom.String()))
-			},
-				Entry("2.332", "2.332"),
-				Entry("1.234", "1.234"),
-				Entry("1.0", "1.0"),
-			)
-
-		})
 		Context("with configmap in VMI annotations for sidecar", func() {
 			var vmi *v1.VirtualMachineInstance
 
@@ -5518,7 +5487,8 @@ var _ = Describe("Template", func() {
 					Expect(err).ToNot(HaveOccurred())
 					Expect(pod.Spec.Containers[0].Name).To(Equal("compute"))
 					expectedMemory := resource.NewScaledQuantity(0, resource.Kilo)
-					expectedMemory.Add(GetMemoryOverhead(&vmi, defaultArch, config.GetConfig().AdditionalGuestMemoryOverheadRatio))
+					launcherRenderer := hypervisor.NewLauncherResourceRenderer(config.GetHypervisor().Name)
+					expectedMemory.Add(launcherRenderer.GetMemoryOverhead(&vmi, defaultArch, config.GetConfig().AdditionalGuestMemoryOverheadRatio))
 					expectedMemory.Add(guestMemory)
 					Expect(pod.Spec.Containers[0].Resources.Requests.Memory().Value()).To(BeEquivalentTo(expectedMemory.Value()))
 					Expect(pod.Spec.Containers[0].Resources.Limits.Memory().Value()).To(BeEquivalentTo(expectedMemory.Value()))
