+++ tests/BUILD.bazel
@@ -96,6 +96,8 @@ go_test(
         "//pkg/hooks/v1alpha1:go_default_library",
         "//pkg/hooks/v1alpha2:go_default_library",
         "//pkg/hooks/v1alpha3:go_default_library",
+        "//pkg/hypervisor:go_default_library",
+        "//pkg/hypervisor/kvm:go_default_library",
         "//pkg/libdv:go_default_library",
         "//pkg/libvmi:go_default_library",
         "//pkg/libvmi/cloudinit:go_default_library",
